:truck: [Vitest] Move test config folder
