+++ vite.config.ts
@@ -15,7 +15,7 @@ const DTS_ENTRY_CONTENT = `export * from './types/index';`;
 const testConfig: vitestTypes.InlineConfig = {
   global: true,
   environment: 'jsdom',
-  setupFiles: './src/test/setup.ts',
+  setupFiles: './test/setup.ts',
 };
 
 // TODO: Do we actually need to include React DOM?
